[qt] Move DataLibrarianWindow and publish dialogs from mktdata to refdata
DataLibrarianWindow is a reference data catalogue browser; its presence in
ores.qt.mktdata forced the Qt MOC-generated staticMetaObject symbols
(ClientCatalogModel, ClientDatasetDependencyModel, etc.) to cross the DLL
boundary on Windows, causing linker errors.

Move DataLibrarianWindow, PublishDatasetsDialog, PublishBundleWizard and
PublicationHistoryDialog to ores.qt.refdata where all their model
dependencies already live. Register the Data Librarian under the
Reference Data > Data Catalogue submenu in RefdataPlugin.
